backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
langchain==0.1.8
httpx==0.26.0
websockets==12.0
pytest==8.0.1
alembic==1.13.1 